requirements.txt
gunicorn
whitenoise
Django==3.2
scikit-learn==1.2.1
pandas==1.3.5

pandas_ta==0.3.14b0
numpy==1.21.5
yfinance
metaapi-cloud-sdk==21.1.0
APScheduler==3.6.3
python-telegram-bot==13.10
psycopg2
django_heroku
textblob
nltk